Moves tertiary button to its own component
