+++ lib/ButtonNew/ButtonTertiary/ButtonTertiary.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import cx from 'classnames';
-import { ButtonBase } from './ButtonBase';
-import { propTypes, defaultProps, sizes } from './common';
+import { ButtonBase } from '../ButtonBase';
+import { propTypes, defaultProps, sizes } from '../common';
 
 function ButtonTertiary({ children, className, disabled, danger, ...rest }) {
   const classes = cx('bg-white', className, {
