+++ lib/ButtonNew/stories/ButtonTertiary/ButtonTertiary.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { map } from 'lodash';
-import { ButtonTertiary } from '../ButtonTertiary';
-import StoryItem from '../../../stories/styleguide/StoryItem';
-import { ButtonStoryItem } from './ButtonStoryItem';
+import { ButtonTertiary } from '../../ButtonTertiary/ButtonTertiary';
+import StoryItem from '../../../../stories/styleguide/StoryItem';
+import { ButtonStoryItem } from '../ButtonStoryItem';
 
 storiesOf('Components', module).add('ButtonTertiary', () => (
   <div>
