+++ lib/index.js
@@ -125,4 +125,4 @@ export { ButtonPrimary } from './ButtonNew/ButtonPrimary/ButtonPrimary';
 export {
   ButtonPrimaryDanger
 } from './ButtonNew/ButtonPrimary/ButtonPrimaryDanger';
-export { ButtonTertiary } from './ButtonNew/ButtonTertiary';
+export { ButtonTertiary } from './ButtonNew/ButtonTertiary/ButtonTertiary';
