+++ stories/index.js
@@ -69,4 +69,4 @@ import { PillInputStory } from '../lib/PillInput/stories/story';
 import DismissiblePrompt from '../lib/DismissiblePrompt/stories/DismissiblePromptStory';
 import '../lib/ButtonNew/stories/ButtonPrimary/ButtonPrimary'
 import '../lib/ButtonNew/stories/ButtonPrimary/ButtonPrimaryDanger'
-import '../lib/ButtonNew/stories/ButtonTertiary'
+import '../lib/ButtonNew/stories/ButtonTertiary/ButtonTertiary'
